Create a class for Date
